added search for task function

+++ src/scripts/task.js
@@ -1,12 +1,15 @@
 //functions for frontend access to tasks collecion (firestore)
 
 import { db, auth } from "../config/firebase";
-import { 
+import {
     addDoc,
     collection,
     getDocs,
     Timestamp,
-    deleteDoc
+    deleteDoc,
+    query,
+    where,
+    doc,
 } from "@firebase/firestore";
 
 //adds task to firestore
@@ -35,6 +38,9 @@ export const addTaskToFirestore = async (taskData) => {
 export const fetchTasksFromFirestore = async () => {
     try {
         const user = auth.currentUser;
+        if (!user) {
+            throw new Error("User not authenticated.");
+        }
 
         const userTasksRef = collection(db, `tasks/${user.uid}/created`);
         const querySnapshot = await getDocs(userTasksRef);
@@ -51,9 +57,13 @@ export const fetchTasksFromFirestore = async () => {
 };
 
 
+//deletes a task from its id
 export const deleteTaskFromFirestore = async (taskId) => {
     try {
         const user = auth.currentUser;
+        if (!user) {
+            throw new Error("User not authenticated.");
+        }
 
         const taskToDelete = collection(db, `tasks/${user.uid}/created/${taskId}`);
         const result = await deleteDoc(taskToDelete);
@@ -61,6 +71,37 @@ export const deleteTaskFromFirestore = async (taskId) => {
         return result;
     } catch (error) {
         console.error("Error deleting task: ", error);
+        throw error;
+    }
+}
+
+
+//search for task from an input
+export const searchForTask = async (inputQuery) => {
+    try {
+        const user = auth.currentUser;
+        if (!user) {
+            throw new Error("User not authenticated.");
+        }
+
+        const tasksRef = collection(db, `tasks/${user.uid}/created/`);
+        const q = query(tasksRef, where("name", "==", inputQuery));
+
+        const querySnapshot = await getDocs(q);
+
+        if (querySnapshot.empty) {
+            console.log("No matching tasks found.");
+            return [];
+        }
+
+        const searchedTasks = querySnapshot.docs.map((doc) => ({
+            id: doc.id,
+            ...doc.data(),
+        }));
+
+        return searchedTasks;
+    } catch (error) {
+        console.error("Error searching for task: ", error);
         throw error;
     }
 }
